refactor(datasets): consolidate converter into command module
