+++ lalamo/evals/datasets/__init__.py
@@ -1,5 +1,5 @@
 from lalamo.evals.datasets.callbacks import BaseConversionCallbacks
-from lalamo.evals.datasets.converter import convert_dataset
+from lalamo.evals.datasets.command import convert_dataset
 from lalamo.evals.datasets.specs import REPO_TO_EVAL, EvalSpec
 
 __all__ = [
